LayoutTests/accessibility/ios-simulator/accessibility-make-first-responder.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/accessibility-helper.js"></script>
<script src="../../resources/js-test.js"></script>
<script>
description("Test that focusing an AX object does not result in web content taking first-respondership.");
window.jsTestIsAsync = true;

function performTest() {
    if (!window.testRunner)
        return;

    setTimeout(async function() {
        debug("Resigning web content first-respondership");
        await resignWebContentFirstResponder();
        await verifyWebContentIsFirstResponder(false);

        var inputObj = accessibilityController.accessibleElementById("input");
        debug("Performing accessibility object focus. Web content should not become the first responder.");
        inputObj.takeFocus();

        // Wait for IPC to occur.
        await sleep(200);
        await verifyWebContentIsFirstResponder(false);

        finishJSTest();
    }, 0);
}

var webContentIsFirstResponder;
async function verifyWebContentIsFirstResponder(expectFirstRespondership) {
    // Set the global var to the opposite of what we expect and allow the UI script to give us the expected value.
    webContentIsFirstResponder = !expectFirstRespondership;

    // Because taking and resigning first respondership requires async IPC between the web and UI processes, we
    // must also asynchronously wait to get our expected value.
    await waitFor(() => {
        setTimeout(async () => {
            webContentIsFirstResponder = await (new Promise(resolve => {
                testRunner.runUIScript(`uiController.webContentIsFirstResponder`, result => resolve(result === "true"));
            }));
        }, 0)
        return webContentIsFirstResponder == expectFirstRespondership;
    });
    shouldBe("webContentIsFirstResponder", `${expectFirstRespondership}`);
}

async function resignWebContentFirstResponder() {
    await (new Promise(resolve => {
        testRunner.runUIScript(`uiController.resignFirstResponder()`, resolve);
    }));
}

window.addEventListener('load', performTest, false);
</script>
</head>
<body>
    <input id="input" type="text">
</body>
</html>
